Fix ref tag case sensitivity
At the moment it's not possible to correctly use `ref:` tag names
which contain uppercase characters (I'm not a huge fan of them but
it has become common to namespace `ref:` tags with the uppercase
country code).

Overpass is case sensitive but osm_conflate currently lowercases all tag
names. This change uses the CaseInsensitiveDict class from `requests`
(which is already a dependency) in SourcePoint to avoid tag name case
issues.